third_party/WebKit/LayoutTests/fast/dom/shadow/get-destination-insertion-points-shadow-insertion-points-expected.txt

CONSOLE WARNING: line 87: Calling Element.createShadowRoot() for an element which already hosts a shadow root is deprecated and will be removed in M63, around December 2017. See https://www.chromestatus.com/features/4668884095336448 for more details.
Tests for getDestinationInsertionPoints() which involves a shadow insertion point.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS nodeList.length is expectedNodes.length
PASS nodeList.item(0) is expectedNodes[0]
PASS successfullyParsed is true

TEST COMPLETE
